agent_backend/requirements.txt
fastapi
uvicorn
langchain
pydantic
mcp
langchain-google-genai
python-dotenv
